Document Ruggeduino custom firmware for pyenv
This is a document explaining to teams how to communicate with the
Ruggeduino in their code using pyenv when it has a custom firmware.

It creates a page under the ruggeduinos page in the programming
documentation, and is split into sections explaining how this is done
with completely custom firmware and with modified SR firmware.