ruggeduino: Swap fully-custom and extended fw secs
I believe that the most common use-case for a customised firmware on
the Ruggeduino will be a team that wants to connect some kind of
sensor with an interface like SPI or I2C.  This type of scenario fits
well with the extended-SR firmware option, and so this should feature
first in the documentation.

This patch swaps the completely-custom and extended firmware
sections.

Change-Id: Ibb725d13a6848414cd3c8b13c574fc41ce4b2519