Hide seconds in times on the home page.
This level of detail isn't needed here, so move it to a title instead.
This involves changing the internal storage of the match times to a
Date object, which is then rendered by AngularJS via changes in the
templates. It also gives us greater freedom to change this more
easily in the future.

Change-Id: I4a4fb20aaafe4f9fb878a579d92bac9a54829e46